pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "runway-lane-detection"
version = "0.1.0"
description = "Awesome `runway-lane-detection` project!"
readme = "README.md"
authors = ["Awesome user <awesome-user@example.com>"]
repository = ""
homepage = ""

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = ["detection"]

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  #! Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
]

[tool.poetry.dependencies]
python = "~3.9.0"
rich = "^10.7.0"
python-dotenv = "^0.21.1"
checksumdir = "^1.2.0"
pandas = "^2.0.2"
opencv-python-headless = "^4.7.0.72"
numpy = "^1.24.3"
shapely = "^2.0.1"
scikit-learn = "^1.2.2"
tqdm = "^4.65.0"
imageio = "^2.31.0"
imageio-ffmpeg = "^0.4.8"
typer = "^0.9.0"
click = "^8.1.3"
project-paths = "^1.1.1"
matplotlib = "^3.7.1"
torch = {version = ">=1.13,<2.0", platform = "linux"}
torchvision = {version = ">=0.14,<1.0", platform = "linux"}
seaborn = "^0.12.2"
ipywidgets = "^8.0.6"

[tool.poetry.group.dev.dependencies]
isort = {extras = ["colors"], version = "^5.10.1"}
nbdime = "^3.1.0"
envyaml = "^1.9.210927"
ipykernel = "^6.23.1"
wemake-python-styleguide = "^0.18.0"
pre-commit = "^2.14.0"
matplotx = "^0.3.10"

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 39
line_length = 100

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]

[tool.project-paths]
readme = "README.md"
data = "data"
configs = "configs"
yolo_dpath = "runway_lane_detection/yolov5-7.0"
